pyproject.toml
[project]
name = "trading-network-for-electronics"
version = "0.1.0"
description = ""
authors = [
    { name = "Dr.T.", email = "Aohudyakov@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "django-phonenumber-field (>=8.3.0,<9.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-select2 (>=8.4.3,<9.0.0)",
    "redis (>=7.0.1,<8.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-filter (>=25.2,<26.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "phonenumbers (>=9.0.18,<10.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
mypy = "^1.18.2"
flake8 = "^7.3.0"
black = "^25.9.0"
isort = "^7.0.0"


[tool.poetry.group.dev.dependencies]
pillow = "^12.0.0"
dotenv = "^0.9.9"
autopep8 = "^2.3.2"
phonenumbers = "^9.0.17"
ipython = "^9.7.0"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
multi_line_output = 3
